Refactored build into run-build.  Having cli/build + cli/e2e call it.
